requirements.txt
ccxt
pandas
flask
python-telegram-bot
matplotlib
